Make RANDP_NUM_BYTES an int
rand_bytes_remaining is an int, so a size_t RANDP_NUM_BYTES narrowed on
assignment.  Casting sizeof keeps the pool size, the assignment, and the
pool index all in int, and gives randp.c and randp.hpp the same type.

+++ benchmarks/randp.hpp
@@ -54,7 +54,7 @@ struct randp
 {
     static_assert(RANDP_NUM_BLOCKS >= 1, "randp must have at least 1 block");
 
-    static constexpr int RANDP_NUM_BYTES = RANDP_NUM_BLOCKS * sizeof(__m128i);
+    static constexpr int RANDP_NUM_BYTES = RANDP_NUM_BLOCKS * (int)sizeof(__m128i);
 
     static_assert(RANDP_RESEED_INTERVAL >= 1, "randp reseed interval must be positive");
